pyproject.toml
[project]
name = "legal-ai-assistant"
version = "0.1.0"
description = "Contract clause extraction, risk scoring, comparison, and reporting"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.40.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pdfplumber>=0.11.0",
  "python-docx>=1.1.0",
  "tiktoken>=0.7.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

